school management sysytem
This is a school management portal system which can be used to add, edit and delete records of students and employee. The code is in c++ . After complete execution and the code running successfully, a file is created in the project folder which has complete information of the data we have entered into the system.